rename: cacp → cacp-protocol (PyPI conflict) + add release.yml workflow
The unprefixed `cacp` slot on PyPI is occupied by an unrelated project
(Classification Algorithms Comparison Pipeline by Sylwester Czmil,
v1.0.4). To publish the reference parser without colliding:

- Distribution name: `cacp-protocol`
- Import name: `cacp_protocol`

Pure mechanical rename — no behavioral change. 42 tests pass.

Files renamed:
- src/cacp/{__init__,models,parser}.py → src/cacp_protocol/...
- All imports updated: `from cacp.X` → `from cacp_protocol.X`,
  `from cacp import` → `from cacp_protocol import`

`pyproject.toml` updates `name = "cacp-protocol"`. README install +
usage examples updated; added a "Naming" callout explaining the
collision so future readers don't wonder.

Added `.github/workflows/release.yml`:
- Triggers on `v*` tag push
- Verifies pyproject version matches the tag
- Builds sdist + wheel via `python -m build`
- Publishes via `pypa/gh-action-pypi-publish` using PyPI
  trusted-publisher (OIDC, no token needed once configured on PyPI)
- Fallback path documented in the workflow comments — operator can
  swap to `password: \${{ secrets.PYPI_API_TOKEN }}` if trusted-publisher
  isn't viable

Operator next steps to ship v0.1.0:
1. Merge this PR
2. (one-time) On pypi.org: register `cacp-protocol` project (it's
   currently unclaimed, HTTP 404 on /pypi/cacp-protocol/json) AND
   add a trusted-publisher entry pointing at owner=zenprocess,
   repo=cacp, workflow=release.yml, environment=pypi
3. Tag `git tag v0.1.0 && git push --tags` — workflow auto-publishes

Unblocks switchyard CACP consolidation epic (#1015), spec 110 chain
(#1017 A/B telemetry already shipped in switchyard v5.4.10).
