+++ .github/workflows/release.yml
@@ -1,0 +1,70 @@
+name: release
+
+# Triggers PyPI publish when a tag matching `v*` is pushed.
+# Uses PyPI trusted-publisher (OIDC) so no secrets need to live in
+# this repo. Setup, one-time, on PyPI side:
+#   pypi.org → manage cacp-protocol → publishing → add trusted publisher
+#     owner: zenprocess
+#     repository: cacp
+#     workflow: release.yml
+#     environment: pypi
+#
+# Fallback: if trusted-publisher is not configured, the upload step
+# fails with a loud "trusted publisher not registered" error. Add
+# `PYPI_API_TOKEN` as a repo secret and replace the `with:` block on
+# the publish step with `password: ${{ secrets.PYPI_API_TOKEN }}` to
+# unblock without OIDC.
+
+on:
+  push:
+    tags:
+      - "v*"
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    name: Build sdist + wheel
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: cacp-python
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+      - name: Install build tooling
+        run: python -m pip install --upgrade pip build
+      - name: Build distributions
+        run: python -m build
+      - name: Verify version matches the tag
+        run: |
+          set -eu
+          DIST_VERSION=$(python -c "import tomllib; print(tomllib.load(open('pyproject.toml','rb'))['project']['version'])")
+          TAG_VERSION="${GITHUB_REF_NAME#v}"
+          if [ "$DIST_VERSION" != "$TAG_VERSION" ]; then
+            echo "::error::pyproject version $DIST_VERSION does not match tag v$TAG_VERSION"
+            exit 1
+          fi
+      - uses: actions/upload-artifact@v4
+        with:
+          name: dist
+          path: cacp-python/dist/
+
+  publish:
+    name: Publish to PyPI
+    needs: build
+    runs-on: ubuntu-latest
+    environment: pypi
+    permissions:
+      id-token: write  # required for trusted-publisher OIDC
+    steps:
+      - uses: actions/download-artifact@v4
+        with:
+          name: dist
+          path: dist/
+      - uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          packages-dir: dist/
