+++ cacp-python/src/cacp_protocol/models.py
